Implement weighted geo_shape centroid support (#50297)
This PR implements proper centroid calculations of geometries according to the
definition defined in #49887. To compute things correctly, an additional variable
encoded long representing the total weight for the centroid of the geometry in
a tree. This weight is always positive. Some tests are fixed, as they did not
have valid geometries.

closes #49887.